Add files via upload
Currently this Dairy daze app can take only take as input as input from use but now I am working on it to take images along side with text and I am in a way to improve UI and making it more beautiful. 

This app uses flutter , SQL and to upload images I am going to use fire base as well.